perf(homepage): extract lib build assets and lazy-load videos
Prevent Vite lib mode from inlining large media into index.js via
?no-inline imports and emitAssets. Defer demo video loading until
sections enter the viewport with useLazyVideo.
